Добавление и удаление знаний

+++ handlers/intent.py
@@ -11,7 +11,7 @@ async def do(func, arg_obj):
     executor = ThreadPoolExecutor(max_workers=100)
     return await loop.run_in_executor(executor, func, arg_obj)
 
-def get_data(jsn, login = None):
+def get_data(jsn):
     db = DB()
     dt = db.exec('''
         SELECT
@@ -25,6 +25,52 @@ def get_data(jsn, login = None):
         table_row = {
             "id_intent": row['id_intent'],
             "name_intent": row['name_intent'],
+            "date_insert": row['date_insert']
+        }
+        table.append(table_row)
+    return table
+
+def get_answer(jsn):
+    db = DB()
+    dt = db.exec('''
+        SELECT
+            id_answer,
+            text_answer,
+            date_insert
+        FROM 
+            Answer
+        WHERE 
+            id_intent = {0}
+        ;
+    '''.format(jsn['id_intent']))
+    table = []
+    for row in dt.table:
+        table_row = {
+            "id_answer": row['id_answer'],
+            "text_answer": row['text_answer'],
+            "date_insert": row['date_insert']
+        }
+        table.append(table_row)
+    return table
+
+def get_example(jsn):
+    db = DB()
+    dt = db.exec('''
+        SELECT
+            id_example,
+            text_example,
+            date_insert
+        FROM 
+            Example
+        WHERE 
+            id_intent = {0}
+        ;
+    '''.format(jsn['id_intent']))
+    table = []
+    for row in dt.table:
+        table_row = {
+            "id_example": row['id_example'],
+            "text_example": row['text_example'],
             "date_insert": row['date_insert']
         }
         table.append(table_row)
@@ -47,8 +93,42 @@ def delete(jsn):
     try:
         db = DB()
         dt = db.exec("""
+            DELETE FROM Example where id_intent = {0};
+        """.format(jsn['id_intent']))
+        if dt.err:
+            return False, str(dt.err)
+        dt = db.exec("""
+            DELETE FROM Answer where id_intent = {0};
+        """.format(jsn['id_intent']))
+        if dt.err:
+            return False, str(dt.err)
+        dt = db.exec("""
             DELETE FROM Intent where id_intent = {0};
         """.format(jsn['id_intent']))
+        if dt.err:
+            return False, str(dt.err)
+    except Exception as ee:
+        return False, str(ee)
+    return True, None
+
+def delete_example(jsn):
+    try:
+        db = DB()
+        dt = db.exec("""
+            DELETE FROM Example where id_example = {0};
+        """.format(jsn['id_example']))
+        if dt.err:
+            return False, str(dt.err)
+    except Exception as ee:
+        return False, str(ee)
+    return True, None
+
+def delete_answer(jsn):
+    try:
+        db = DB()
+        dt = db.exec("""
+            DELETE FROM Answer where id_answer = {0};
+        """.format(jsn['id_answer']))
         if dt.err:
             return False, str(dt.err)
     except Exception as ee:
@@ -77,7 +157,25 @@ class Handler:
                     return web.json_response({'result':True, 'err': None, 'table':[]})
             except Exception as ee:
                 return web.json_response({"result":False,"err":str(ee),"table":[]})
-        if method == "add":
+        if method == "get_answer":
+            try:
+                table = await do(get_answer, jsn)
+                if len(table) > 0:
+                    return web.json_response({'result':True, 'err': None, 'table':table})
+                else:
+                    return web.json_response({'result':True, 'err': None, 'table':[]})
+            except Exception as ee:
+                return web.json_response({"result":False,"err":str(ee),"table":[]})
+        if method == "get_example":
+            try:
+                table = await do(get_example, jsn)
+                if len(table) > 0:
+                    return web.json_response({'result':True, 'err': None, 'table':table})
+                else:
+                    return web.json_response({'result':True, 'err': None, 'table':[]})
+            except Exception as ee:
+                return web.json_response({"result":False,"err":str(ee),"table":[]})
+        if method == "add_intent":
             try:
                 table, err = await do(add, jsn)
                 if err:
@@ -95,7 +193,24 @@ class Handler:
                     return web.json_response({"result":False,"err": str(err)})
             except Exception as ee:
                 return web.json_response({"result":False,"err":str(ee)})
-
+        if method == "delete_example":
+            try:
+                result, err = await do(delete_example, jsn)
+                if result:
+                    return web.json_response({"result":True, "err": None})
+                else:
+                    return web.json_response({"result":False,"err": str(err)})
+            except Exception as ee:
+                return web.json_response({"result":False,"err":str(ee)})
+        if method == "delete_answer":
+            try:
+                result, err = await do(delete_answer, jsn)
+                if result:
+                    return web.json_response({"result":True, "err": None})
+                else:
+                    return web.json_response({"result":False,"err": str(err)})
+            except Exception as ee:
+                return web.json_response({"result":False,"err":str(ee)})
 
         return web.json_response({"result": False, "err": "Метод не найден", "data": None})
